Fix: Corrige le routage et l'import des fichiers CSS
- Centralise le routage dans App.jsx avec createBrowserRouter et RouterProvider
- Corrige les chemins d'import des composants (SignUp, LogIn)
- Supprime l'import incorrect de App.css dans SignUp.jsx
- S'assure que le CSS global est importé uniquement dans App.jsx
- Nettoie main.jsx pour ne garder que le rendu de <App />
