--- src/components/SignUp.jsx
@@ -1,11 +1,0 @@
-import '../App.css';
-
-
-
-const SignUp = () => {
-return (
-    <div className='text-white-500'>Hello world</div>
-)
-}
-
-export default SignUp
