+++ src/components/auth/SignUp.jsx
@@ -1,0 +1,5 @@
+const SignUp = () => {
+  return <div className="text-white-500">Hello world</div>;
+};
+
+export default SignUp;
